back-end/requirements.txt
PyYAML==5.3.1
selenium==3.141.0
Flask-RESTful==0.3.8
Flask==1.1.2
Flask-Cors==3.0.9